refactor(core): move CostTrack to the core.usage leaf
`clients/openrouter.py` -- an HTTP client -- imported `CostTrack` from
`core/benchmark.py`, a 1.7k-line benchmark harness, to build one in
`make_token_cost_track`: the floor importing the ceiling. Both statements were
lazy (one TYPE_CHECKING, one function-local), so the runtime graph never saw the
edge while grimp/import-linter did -- and that single edge held a 9-module
all-edges SCC together.

`CostTrack` + its `CostBasis` alias now live in `core/usage.py`, beside the
`LLMUsage` that `CostTrack.usage` holds: tokens are the fact, dollars are
derived, so both models on that sentence belong together, and `core.usage`
imports nothing from langres. Pure move -- no signature or field changed.

`benchmark._cost_track` deliberately stays put: it reads the harness's
`PairwiseJudgement.provenance` conventions, so moving it would drag
`core.models` in and cost `core.usage` the leaf property that makes it a safe
home.

Measured (uv run python tools/import_graph.py kinds):
  all-edges SCCs [9, 9, 3, 2] -> [9, 3, 2, 2, 2], tangled 23 -> 18
  runtime SCCs   [] -> []      (unchanged: the edge was never toplevel)

The litellm-seam 9 did not shrink, it disintegrated into {anchor_store,
resolver} and {benchmark, methods}. Five modules left the tangle:
clients.openrouter, core.finetune, core.matchers.cascade,
core.matchers.llm_judge, core.method_registry. largest_scc stays 9 -- that is
now the untouched data.data_profile.* knot, a different component of the same
size.
